chore:재생목록 api 파일 변경
